Great channel improvements.
The old receive feature is still there, but there's a new faster one not
depending on exceptions, but rather allows you to specify two callbacks,
one for data (if such ever comes), and one for a close event.

The send() interface has changed, where it used to throw a
channel_closed exception if the channel was closed (i.e. the message
wasn't sent). Now it returns false instead (and true if the message was
successfully sent). ensure_send() is available, and works like send()
used to work.