readable::consume() now handles concurrency
This means that a second optional argument can be provided specifying
the maximum concurrency which should be allowed. The callback
function will be called sequentially, but its returning promise will
not have to be awaited before running the callback again - instead a
new callback can be called with the next item, until at most n
outstanding not-yet-resolved promises are in motion. When one of them
resolves, the callback can be called again with the next item in the
channel (given there is one).