q-test refactoring WRT gtest and CMake
q-test is now (on modern CMake's) a header-only library without
an explicit dependency on gtest, but an implicit "peer dependency".
It's up to the target unit test to actually link to gtest.

It's also prepared for more backends and not just gtest.

It builds properly with CMake, allowing easy integration with end
application the same way as @jupp0r's PR for q itself.